redirect on headersReceived.
That we can clearly see the redirection in devtools, with appropriate host.

NOTE: redirect is always done with GET method.
Should not be an issue, intent is to redirect a site/widget entirely, using
its own urls path, so not redirected; other methods are OK. That would be
an issue only when you redirect specific non-GET api methods. Which would
be strange and quite prone to errors down the road.